create upstream module. move rpc into upstream
